{"schema":"decodex/commit/1","summary":"Split promotion payload modules","authority":"manual"}
